+++ src/main/java/controller/DeleteQcmServlet.java
@@ -12,13 +12,18 @@ public class DeleteQcmServlet extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws IOException {
 
-        int num = Integer.parseInt(request.getParameter("num_question")) ;
+        // Vérification session admin
+        HttpSession session = request.getSession();
+        if (session.getAttribute("admin") == null) {
+            response.sendRedirect("AdminPage.jsp");
+            return;
+        }
+
+        int num = Integer.parseInt(request.getParameter("num_question"));
 
         try {
             QcmDAO dao = new QcmDAO();
             dao.delete(num);
-            System.out.println("Num = [" + num + "]");
-
             response.sendRedirect("ListeQcmServlet");
 
         } catch (Exception e) {
